Add files via upload
First commit of the new Project Owl ClusterDuck-Protocol check and
installation script for Linux.  It will check to make sure that specific
Linux binary applications are avalible and if not ask the user to install
them.  Then it checks for the Arduino IDE and finally it can install the
Project Owl ClusterDuck-Protocol libraries.
 
There is more work to be done, the .arduino15/preferences.txt still needs
editing to put the information in the boardsmanager.additional.urls setup. 
That will take sed or awk not sure.  And there is a test that needs to be
written looking to make sure the sub libraries actually got installed.
    
Finally the libraries in Arduino/libraries/ClusterDuck-Protocol/Libraries/
are now symlinked back to Arduino/libraries/ instead of being copied or
moved there.  git will still work to keep everything updated.